pyproject.toml
[tool.poetry]
name = "cognoscere"
version = "0.1.0"
description = "Aplikacja wspierająca psychoterapię poprzez ustrukturyzowany dialog terapeutyczny między sesjami"
authors = ["Cognoscere Team"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { version = "^0.115.0", extras = ["standard"] }
sqlalchemy = { version = "^2.0.0", extras = ["asyncio"] }
asyncpg = "^0.29.0"
alembic = "^1.13.0"
pydantic-settings = "^2.7.0"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
bcrypt = "^4.2.0"
openai = "^1.50.0"
python-multipart = "^0.0.12"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"
ruff = "^0.9.0"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
